Fix for issue #125.
Kratu Dates are now handled at time 12:00 instead of 0:00.  This
replicates the format used in AW Reports and resolves this issue for all
timezones.

Existing Kratus in the Kratu database table needs to be updated to the
new time.  This can be by executing the following in MySQL:

mysql> update AW_Kratu set DAY = DATE_ADD(DAY, INTERVAL 12 hour) where
HOUR(DAY) IN(0,1);